mailbox: updato to HW Spec v2018-10-22
In this HW spec, there are 16 interrupts from each IP block, and each of
the two events from each instance can be mapped to any of the 16
interrupts.

In our implementation, we use one pair of interrupts per subsystem, and
map events from all mailboxes that the subsystem cares about onto that
subsystem's pair of interrupts.