wdt: subsystem is allowed to enable its own timer
The HW spec does not explictly state this, but it does not explicitly
deny this either, so this interpretation is consistent with the spec.
And, it certainly makes the SW simpler, since the subsystem can
enable the WDT only once it's ready to start kicking it.

TRCH becomes responsible for implementing a timeout on boot time, i.e.
the interval between the reset command from TRCH and the enabling of the
WDT by the subsystem.