make: strip ELF binary
Stripped debug symbols, and unnecessary alignment gaps.

The ELF binary is now almost the same size as the .bin file,
but has the advantage of having an ELF header. This is particularly
useful for rtps.elf, which can now be loaded by u-boot with bootelf
(and the rtps.elf would no longer be wasting NV mem).