trch: subsys: add concept of a CPU Group
A CPU Group is a lower-level concept than a subsystem. A subsystem
abastracts the details (e.g. RTPS subsystem may be booted as one OS in
LOCKSTEP, or as two OSes in SPLIT, or as one OS in SMP).

Most code should be using CPU Group.  Only the very top-level code might
want to use the subsystem concept, if it desires to present its user
with a very high-level notion of a subsystem.