gtimer: fix timer freq to 125MHz
Generic Timer increments by 1 on every update of Elapsed Timer.
Elapsed Timer updates at 125MHz clock, but increments by 8 each
time. So, max value of Generic Timer is 1/8th of the max value of
Elapsed Timer, i.e. Generic Timer is three bits narrower than
Elapsed Timer.

The 1MHz that was there was wrong (and so was the busyloop factor,
because the two were calibrated to each other).

Also, move to hwinfo.