plat: smc: revert to single SMC-353 (memmap v14)
The change to dual SMCs is a change that's part of Chiplet Mem Map v16.
But, v16 is not yet supposed to be in snap (nor in zebu) branch, but
by mistake, this one part of v16 was merged into snap.

Mem Map v14 has a total of 128 MB allocated to SRAM/NAND, so
make SRAM 16MB (x4) and NAND 64MB.