rtps: SMP mode
Make it a mode in which the application is run, instead of a standalone test.
Standalone tests should be side-effect free; and also, this opens the
more functionality in the app related to SMP, e.g.  processing interrupts on
both cores (which would already happen, if Qemu model would round-robin
among cores in 1-of-N mode, but it does not).